requirements.txt
Django==5.2.4
Pillow==10.1.0 
setuptools>=65.5.0  
django-crispy-forms==2.1
crispy-bootstrap5==0.7
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7
django-debug-toolbar==4.2.0

